+++ src/ReturnTypes/Helpers/AppExtension.php
@@ -5,24 +5,20 @@ declare(strict_types=1);
 namespace NunoMaduro\Larastan\ReturnTypes\Helpers;
 
 use Illuminate\Foundation\Application;
-use NunoMaduro\Larastan\Concerns\HasContainer;
-use PhpParser\Node\Expr;
-use PhpParser\Node\Expr\ClassConstFetch;
+use NunoMaduro\Larastan\ReturnTypes\AppMakeHelper;
 use PhpParser\Node\Expr\FuncCall;
-use PhpParser\Node\Name\FullyQualified;
-use PhpParser\Node\Scalar\String_;
 use PHPStan\Analyser\Scope;
 use PHPStan\Reflection\FunctionReflection;
 use PHPStan\Type\DynamicFunctionReturnTypeExtension;
-use PHPStan\Type\ErrorType;
-use PHPStan\Type\NeverType;
 use PHPStan\Type\ObjectType;
 use PHPStan\Type\Type;
-use Throwable;
 
 class AppExtension implements DynamicFunctionReturnTypeExtension
 {
-    use HasContainer;
+    public function __construct(
+        private AppMakeHelper $appMakeHelper
+    ) {
+    }
 
     public function isFunctionSupported(FunctionReflection $functionReflection): bool
     {
@@ -38,28 +34,6 @@ class AppExtension implements DynamicFunctionReturnTypeExtension
             return new ObjectType(Application::class);
         }
 
-        /** @var Expr $expr */
-        $expr = $functionCall->getArgs()[0]->value;
-
-        if ($expr instanceof String_) {
-            try {
-                /** @var object|null $resolved */
-                $resolved = $this->resolve($expr->value);
-
-                if ($resolved === null) {
-                    return new ErrorType();
-                }
-
-                return new ObjectType(get_class($resolved));
-            } catch (Throwable $exception) {
-                return new ErrorType();
-            }
-        }
-
-        if ($expr instanceof ClassConstFetch && $expr->class instanceof FullyQualified) {
-            return new ObjectType($expr->class->toString());
-        }
-
-        return new NeverType();
+        return $this->appMakeHelper->resolveTypeFromCall($functionCall);
     }
 }